Extract `verify_system_bundle_solution()` in sp-domains
No logical changes.
